Enable linting
Fix linting issues

Move components into components folder

Mess around with webpack.config
