+++ index.html
@@ -1,6 +1,7 @@
 <!doctype html>
 <html>
   <head>
+    <script src="http://localhost:8097"></script>
     <meta charset="UTF-8">
     <title>TTV4TV</title>
     <link rel="stylesheet" href="style.css">
