+++ src/components/search-bar.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default class SearchBar extends React.Component {
+  render () {
+    return null // TODO: implement search bar
+  }
+}
