+++ src/components/side-bar.jsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default class SideBar extends React.Component {
+  render () {
+    return null // TODO: implement side bar
+  }
+}
