+++ webpack.config.babel.js
@@ -1,39 +1,45 @@
-const path = require('path');
-const webpack = require('webpack');
+const path = require('path')
+const webpack = require('webpack')
 
-const sourcePath = path.join(__dirname, 'src');
+const sourcePath = path.join(__dirname, 'src')
 
 const config = {
   entry: ['babel-polyfill', path.resolve(sourcePath, 'app.jsx')],
-  output: { path: __dirname, filename: 'bundle.js', },
+  output: { path: __dirname, filename: 'bundle.js' },
   resolve: {
     extensions: ['.js', '.jsx'],
-    modules: [sourcePath, path.resolve(__dirname, 'node_modules')],
+    modules: [sourcePath, path.resolve(__dirname, 'node_modules')]
   },
   module: {
     rules: [
       {
         test: /\.(js|jsx)$/,
         exclude: /node_modules/,
-        use: ['babel-loader'],
+        loader: 'babel-loader',
         include: sourcePath,
-      },
-    ],
+        options: {
+          presets: ['react']
+        }
+      }
+    ]
   },
-  plugins: [],
-};
+  devtool: 'source-map',
+  context: __dirname,
+  target: 'web',
+  plugins: []
+}
 
 if (process.env.NODE_ENV === 'production') {
   config.plugins.push(
     new webpack.optimize.UglifyJsPlugin()
-  );
+  )
   config.plugins.push(
     new webpack.DefinePlugin({
-      'process.env.NODE_ENV': JSON.stringify('production'),
+      'process.env.NODE_ENV': JSON.stringify('production')
     })
-  );
-  config.plugins.push(new webpack.optimize.ModuleConcatenationPlugin());
-  config.plugins.push(new webpack.HashedModuleIdsPlugin());
+  )
+  config.plugins.push(new webpack.optimize.ModuleConcatenationPlugin())
+  config.plugins.push(new webpack.HashedModuleIdsPlugin())
 }
 
-module.exports = config;
+module.exports = config
